(numba/dppy) Add option to dump intermediate IR files

+++ dppy_driver/spirv_generator.py
@@ -8,6 +8,9 @@ import tempfile
 
 from numba import config
 
+SAVE_DPPY_IR_FILES = 0
+SAVE_DPPY_IR_FILES = environ.get(SAVE_DPPY_IR_FILES)
+
 def _raise_bad_env_path(msg, path, extra=None):
     error_message = msg.format(path)
     if extra is not None:
@@ -72,14 +75,15 @@ class Module(object):
         self._finalized = False
 
     def __del__(self):
-        self.close()
-
-    def close(self):
         # Remove all temporary files
         for afile in self._tempfiles:
-            os.unlink(afile)
+            if SAVE_DPPY_IR_FILES == 1:
+                print(afile)
+            else:
+                os.unlink(afile)
         # Remove directory
-        os.rmdir(self._tmpdir)
+        if SAVE_DPPY_IR_FILES != 1:
+            os.rmdir(self._tmpdir)
 
     def _create_temp_file(self, name, mode='wb'):
         path = self._track_temp_file(name)
@@ -91,9 +95,6 @@ class Module(object):
                             "{0}-{1}".format(len(self._tempfiles), name))
         self._tempfiles.append(path)
         return path
-
-    # def _preprocess(self, llvmir):
-    #    return adapt_llvm_version(llvmir)
 
     def load_llvm(self, llvmir):
         """
